Add GWF conversion script

+++ database/schema.sql
@@ -176,11 +176,11 @@ CREATE TABLE jobs (
 -- A task that's defined in terms of how many flops (floating point operations) it takes to complete
 DROP TABLE IF EXISTS tasks;
 CREATE TABLE tasks (
-  id                 INTEGER PRIMARY KEY     NOT NULL AUTO_INCREMENT,
-  start_tick         INTEGER                 NOT NULL CHECK (start_tick >= 0),
-  total_flop_count   INTEGER                 NOT NULL,
-  job_id             INTEGER                 NOT NULL,
-  parallelizability  VARCHAR(50)             NOT NULL,
+  id                INTEGER PRIMARY KEY     NOT NULL AUTO_INCREMENT,
+  start_tick        INTEGER                 NOT NULL CHECK (start_tick >= 0),
+  total_flop_count  BIGINT                  NOT NULL,
+  job_id            INTEGER                 NOT NULL,
+  parallelizability VARCHAR(50)             NOT NULL,
 
   FOREIGN KEY (job_id) REFERENCES jobs (id)
     ON DELETE CASCADE
